+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add `ExForce.OAuth` to support `authorization_code` and `refresh` in addition to `password` grant type.
+
 ## 0.1.0 - 2017-09-13
 
 ### Added